Home Work 4 done

+++ MetricsAgent/DAL/TimeSpanHandler.cs
@@ -7,7 +7,7 @@ namespace MetricsAgent.DAL
 public class TimeSpanHandler : SqlMapper.TypeHandler<TimeSpan>
     {
         public override TimeSpan Parse(object value)
-            => TimeSpan.FromSeconds((long)value);
+            => TimeSpan.FromSeconds((double)value);
         public override void SetValue(IDbDataParameter parameter, TimeSpan
         value)
         => parameter.Value = value;
